Load segments defaults and extract set global opacity
